[#29763] YSQL: Add yb_disable_pg_snapshot_mgmt_in_repeatable_read GUC for forward compatibility
Summary:
The yb_disable_pg_snapshot_mgmt_in_repeatable_read GUC is added in
2025.1 and 2024.2 to allow a user to disable PG snapshot management for
repeatable read isolation and fallback to the previous mode of operation
in the rare event of an issue. However, this GUC is not added to newer
releases to avoid the code complexity it introduces and the fact that
newer stable releases will have been operating in this mode for longer
to fairly rule out the possiblity of any unknown issues.

Test Plan: Manually verified that upgrading from 2025.1 (with the GUC set) to a newer release with this change works.

Reviewers: smishra, kramanathan

Reviewed By: kramanathan

Subscribers: kramanathan, yql

Differential Revision: https://phorge.dev.yugabyte.com/D49142