[#29705] DocDB: Temporary typedefs for lightweight protobufs migration
Summary:
In order to use lightweight protobufs for read and write requests, we have to change types from SomeTypePB to LWSomeTypePB in many places.
This change involves 200+ files.
To avoid doing it in the diff that actually enables lightweight protobufs and make this diff smaller we could introduce intermediate typedefs SomeTypeMsg, that will be SomeTypePB at this point.
And will be changed to LWSomeTypePB after migration to lightweight protobufs.
It will allow us to decrease size of the diff that actually changes logic.

Test Plan: Jenkins

Reviewers: arybochkin, xCluster, hsunder

Reviewed By: arybochkin

Subscribers: yql, ybase

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D49017