[#29756] YSQL: Interrupt waiting for operation order in case of connection shutdown
Summary:
There are situations when perform operation with particular `serial_no` might be lost. Real world situation is a parallel execution with restart error. In this case leader process will kill the workers and some of their perform operations will not be sent, while `serial_no` will be increased. As a result leader process will send next perform operation with `serial_no` which predecessors will never come.

Shutdown of session in such a state will wait till the predecessor's waiting deadline (which is currently equal to request's deadline and it is `600secs` by default).

This diff interrupts the waiting of predecessors process in case of session shutdown. Also it changes the deadline for predecessors waiting. Now it is min of request deadline and the now + `FLAGS_pg_client_session_expiration_ms` (i.e. the timeout we expect the session's heartbeat to come), The `FLAGS_pg_client_session_expiration_ms` in `60secs` by default.

The diff also fixes long time of backend termination due to waiting for response over shared exchange.

**Note:**
The `RequestSequencer` class uses external mutex from the `LockablePgClientSession` object. And it is expected almost all the nethods be called (except `Shutdown`) under the locked mutex. It is not easy to use thread analysis macros like  `GUARDED_BY` to make sure that mutex is locked in all the cases. As a result alternative runtime approach is used for now

```
  void ProcessPending(PgClientSession& session) {
    DCHECK(DEBUG_IsMutexLocked());
    impl_.ProcessPending(session);
  }

  DEBUG_ONLY([[nodiscard]] bool DEBUG_IsMutexLocked() const {
      return !std::unique_lock(const_cast<std::mutex&>(mutex_), std::try_to_lock).owns_lock();
  });

  [[maybe_unused]] const std::mutex& mutex_;
```

Test Plan:
New unit test are created

```
./yb_build.sh --gtest_filter PgOpProcessingOrderTest.ShutdownTimeWithLostOp
./yb_build.sh --gtest_filter PgOpProcessingOrderTest.BackendTerminationTimeWithLostOp
```

Reviewers: sergei, bkolagani, patnaik.balivada, pjain

Reviewed By: bkolagani

Subscribers: ybase, yql

Tags: #jenkins-trigger, #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D49114