[BACKPORT 2025.2][#28326,#28946] xClusterDDLRepl: Fix missing packing schema issue
Summary:
**Background:**
There is an issue with `InsertPackedSchemaForXClusterTarget` where it can fail on a subset of tablets.
The master may then force update this tablet to the most up to date schema. However, since this rpc
inserts 2 schemas (n+1 is the new schema, n+2 is the current schema that's reinserted), then these
failed tablets would miss the new schema. Thus when pollers try to write to this tablet, they will
fail with a schema not found error.

**Solution:**
To solve this, we need to ensure that every tablet has the inserted schema before continuing with
replication. However, we cannot rely on independent pollers to handle this as they only check their
local tablet - so if there are uneven tablets, its possible that we could miss checking a tablet.

The solution here is to have master handle all of the inserting and verification of the new schema.

For automatic mode, we will now block pollers if they find a new schema that they don't have in
their schema mapping. In such a case they will send a `HandleNewSchemaForAutomaticXClusterTarget`
request to master.
We then maintain this invariant:
> The producer->consumer schema mapping can only be updated once we are 100% sure that all tablets
> have this consumer schema version.

The `HandleNewSchemaForAutomaticXClusterTarget` request will verify that the master's schema mapping
doesn't already include this producer schema. It will then start a
`HandleNewSchemaForAutomaticXClusterTargetTask` for this table. There's some new code added here that
ensures that only one such task can run at once for a table - this helps to reduce any extra work or
extra schema versions inserted.

`HandleNewSchemaForAutomaticXClusterTargetTask` is a multi-step task that follows these steps:
1. Query all tablets for their latest compatible schema version.
   This will send an `AsyncGetCompatibleSchemaVersion` to each tablet with the new producer schema.
2. Check if all tablets have same compatible schema version.
   a. If any tablet does not find a compatible version, or if tablets don't agree on the same
      version then we go to step 3.
   b. If all tablets agree on the same version, then go to step 5.
3. Insert new packed schema.
   Call `InsertPackedSchemaForXClusterTarget` to insert the new schema in to all tablets
4. Wait for `InsertPackedSchemaForXClusterTarget`.
   Wait until the table no longer has any tasks of this type.
   Once completed, we go back to step 1 - this is important to ensure that we catch any failed
   `InsertPackedSchemaForXClusterTarget` tasks
5. Update the producer->consumer schema versions mapping with the agreed upon consumer schema version
   At this point, we are now certain that all tablets have the same compatible schema.
   Once we update the master's mapping, this will update the schema mappings for the pollers, which
   will then unblock these pollers.

This diff also adds additional logic in `HandleNewSchemaForAutomaticXClusterTarget` to handle
upcoming colocated tables. If we find a colocation id that we don't know about (ie hasn't yet been
created), then we will instead call `HandleNewTableSchemaForUpcomingColocatedTable` which will insert
the upcoming schema into the replication group (same as before, but now handled in the master
instead of in the output_client). See D41367 for more info on colocated handling, note that this
is a different from the producer->consumer schema mapping.

**Next steps:**
There are some additional follow-ups that this diff does not address:
- `HandleNewSchemaForAutomaticXClusterTargetTask` needs to verify that the confirmed schema version is
  greater than any other value in the schema mapping. If it is lower, then its possible that this
  schema packing was already gc-ed on a tablet (we gc versions that are lower than the lowest value
  in the mapping). In such a case, `HandleNewSchemaForAutomaticXClusterTargetTask` will have to insert
  a new schema. GHI #29760
- Avoid repeated `InsertPackedSchemaForXClusterTarget` calls. If `InsertPackedSchemaForXClusterTarget`
  keeps failing, this diff will just keep calling it until it gets matching schema versions across
  all tablets. These repeated calls will keep bumping up the schema versions. We should be able to
  capture the first inserted schema version, and then retry inserting the schema at that version
  again on retries. GHI #29761
- Performance improvements: Since we now go to master instead of our local tablet for
  ChangeMetadataOps, this diff introduces some slowness in DDL processing. However much of this is
  required for correctness, so it is necessary.
  There are some potential improvements we could make for single tablet tables / colocated tables.
  Since we only need one tablet to agree upon the value, we could handle this from the poller, but
  this will make the code less readable.

**Upgrade/Rollback safety**
Adding autoflag `xcluster_automatic_mode_schema_handling_v2` that blocks this feature. When false,
we still use the old schema handling (poller checks local tablet). When true, we use the new path.

Original commit: 340ec5318ab7dcdf25dbeedd325e5f677aa121d5 / D49035

Test Plan:
existing tests (this approach slows down some tests a bit more, so needed to bump up the timeouts for tests)

```
ybd --cxx-test integration-tests_xcluster_ddl_replication-test --gtest_filter XClusterDDLReplicationTest.FailAsyncInsertPackedSchema
```

Reviewers: xCluster, hsunder, mlillibridge

Reviewed By: hsunder

Subscribers: ybase

Differential Revision: https://phorge.dev.yugabyte.com/D49164