[#28707] DocDB: Add log line when servicing RestoreSnapshotSchedule
Summary:
The master leader usually logs a line when servicing a major RPC, such as creating/deleting snapshots.
This is useful for debugging incidents. It was missing for the RestoreSnapshotSchedule RPC.
This commit fixes that.

Test Plan:
Existing cxx-test related to RestoreSnapshotSchedule
exercise the code; verified the added logline is present

./yb_build.sh debug --cxx-test yb-admin-snapshot-schedule-test --gtest_filter YbAdminSnapshotScheduleTest.Basic
./yb_build.sh debug --cxx-test snapshot-test --gtest_filter SnapshotTest.RestoreSnapshot
./yb_build.sh debug --cxx-test yb-admin-snapshot-test --gtest_filter AdminCliTest.*RestoreSnapshot*

There are [too] many other tests in the yb-admin-snapshot-schedule-test
exercising the code, but it takes time.  I ran them for sanity w/o any failures/errors.
./yb_build.sh debug --cxx-test yb-admin-snapshot-schedule-test

Reviewers: mhaddad, hsunder

Reviewed By: mhaddad

Subscribers: asrivastava, ybase

Differential Revision: https://phorge.dev.yugabyte.com/D49080