[BACKPORT 2025.2][#29760] xClusterDDLRepl: Validate new consumer schema version with min version in mapping
Summary:
Follow-up to D49035.

This diff verifies that any new consumer schema version that `HandleNewSchemaForAutomaticXClusterTargetTask` adds to the producer->consumer schema versions map is greater than or equal to the minimum consumer version in the map. Since we use the minimum value in the map to know when to stop schema GC (see `FillMinXClusterSchemaVersion` and `GetMinXClusterSchemaVersion`), using a lower schema version has a risk that the schema could have been GC-ed on a tablet in the meantime.

Handling this by adding a `check_min_consumer_schema_version` argument to `UpdateConsumerOnProducerMetadata`. If provided, this will verify that the new consumer version provided is not less than what is currently in the map. If it is too old, then it returns a special error - `HandleNewSchemaForAutomaticXClusterTargetTask` will check for this error, and will instead insert a new schema in such a case.

**Upgrade/Rollback safety:**
Just adding a new error type, not relevant for upgrade/rollback

Original commit: 5753a9f03611dc54febee33c9f50207c5776134f / D49187

Test Plan:
```
ybd --cxx-test integration-tests_xcluster_ddl_replication-test --gtest_filter XClusterDDLReplicationTest.AvoidOldCompatibleConsumerSchemaVersion
```

Reviewers: hsunder, xCluster

Reviewed By: hsunder

Subscribers: xCluster, ybase

Differential Revision: https://phorge.dev.yugabyte.com/D49206