[BACKPORT 2025.1][#29814] YSQL: Attempt to fix flaky test TestWildcardRegionYbAdmin
Summary:
The test errors out 5-10% of the time. Debugging one instance showed that the LB was trying to move leaders of some system tablets (pg_advisory_lock..) and the leader was resisting a change due to the flag min_leader_stepdown_retry_interval_ms, set to 20 secs. Using a smaller value of this flag for this test to check if that fixes it. Also, just allowing more time for the LB to converge in general (30 -> 60s).

Original commit: 9484872 / D49234

Test Plan: Jenkins. Repeated local runs don't reproduce the problem well.

Reviewers: asrivastava, kfranz

Reviewed By: asrivastava

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D50279