[#29821] Docdb: Fix calls to GetDdlTransactionMetadata after #27613
Summary:
The line
```
    const auto* metadata = VERIFY_RESULT(GetDdlTransactionMetadata(
        req.use_transaction(), req.use_regular_transaction_block(), context->GetClientDeadline(),
        req.options().is_using_table_locks()));
```
modified in ff6db4fae3dd44712776af9c012eb3407587dd56 would misinterpret the final
argument as `priority` instead of `txn_using_table_locks.

```
  Result<const TransactionMetadata*> GetDdlTransactionMetadata(
      bool use_transaction, bool use_regular_transaction_block, CoarseTimePoint deadline,
      uint64_t priority = kHighPriTxnUpperBound, uint64_t pg_txn_start_us = 0,
      bool txn_using_table_locks = false)
```

Fixing the function calls to ensure that this is fixed.

Test Plan:
The above code path, would not get executed when table locks are enabled.
This is because when table locks are enabled, DoAcquireObjectLocks will ensure
that the transaction is created, thus not entering the `if (!txn) {` branch in GetDdlTransactionMetadata.

./yb_build.sh release --cxx-test pg_object_locks-test --gtest_filter TableLocksEnabled/PgObjectLocksTestAbortTxns.TestDDLAbortsTxns/*

./yb_build.sh release --cxx-test pg_object_locks-test --gtest_filter TableLocksEnabled/PgObjectLocksTestAbortTxns.TestDDLAbortsTxnsInMixedMode/*

Reviewers: sergei, bkolagani

Reviewed By: sergei

Subscribers: ybase, yql

Differential Revision: https://phorge.dev.yugabyte.com/D49251