[BACKPORT 2.20][PLAT-19283][xCluster] Increase default yb.xcluster.bootstrap_producer_timeout to prevent xCluster setup failures
Summary:
When the number of tables being added to xCluster replication is high, the 2 minutes timer for the BootstrapProducer RPC can prematurely time out. This diff increases this timeout to 20 minutes.

Original commit: D49284

Test Plan: Verified the default value shown for the runtime config value is 20 minutes after the change.

Reviewers: amindrov, spothuraju, jmak, nbhatia

Reviewed By: amindrov, spothuraju

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D49287